refactor: split manuscript_checker.py god-module into focused modules (#118)
Closes #118

Splits ``tools/analysis/manuscript_checker.py`` (2201 LOC, 40 functions
covering 5+ distinct concerns) into a focused package under
``tools/analysis/manuscript/``. The old file becomes a 105-LOC re-export
shim so existing imports keep working without touching any caller.

Package layout
--------------

- vocabularies.py — static wordlists + numeric thresholds (n-gram sizes,
  body parts, blocking verbs, sensory tokens, structural hints, LY
  exclusions, action verbs fallback, question openers, cliché phrases,
  memoir thresholds). Pure data, no functions.
- types.py — Finding + Occurrence dataclasses + classifier
  (``_classify``, ``_looks_structural``).
- text_utils.py — text normalization (``_strip_markdown``, ``_tokenise``,
  ``_strip_dialogue``), n-gram extraction (``_ngrams_in_line``,
  ``_make_snippet``), chapter-draft loader (``_read_chapter_drafts``).
- metadata.py — book/people state readers (``_read_book_genres``,
  ``_read_book_category``, ``_read_people_profiles``,
  ``_read_allowed_repetitions``, ``_read_snapshot_threshold``).
- rules.py — book CLAUDE.md ## Rules parser + scanner
  (``_read_book_rules``, ``_extract_patterns_from_rule``,
  ``_rule_label``, ``_scan_book_rules``).
- scanners.py — fiction-side per-category scanners (filter words,
  adverb density, clichés, question-as-statement, sentence repetitions,
  snapshots, callbacks).
- memoir_patterns.py — five memoir scanners (anonymization leak, tidy
  lesson endings, reflective platitudes, timeline ambiguity, real-people
  consistency).
- renderer.py — ``render_report`` + per-category recommendations.
- __init__.py — ``scan_repetitions`` orchestrator + public re-exports.

Backward compatibility
----------------------

``tools/analysis/manuscript_checker.py`` becomes a thin shim that
re-exports the public API and every private helper the existing tests
and skill-side code import. No caller had to change. Verified:

- 140 tests in test_manuscript_checker.py + test_manuscript_checker_memoir.py
  pass without modification.
- 79 tests in test_hooks.py (which imports the linter helpers indirectly
  via the manuscript_checker shim) pass without modification.
- Cross-module consumers (chapter_validator, banlist loader, server.py
  MCP tool) work unchanged.

Mirror tests (#118 done definition)
-----------------------------------

Adds 57 new tests under ``tests/analysis/manuscript/`` mirroring the
new package structure:

- test_types.py — classifier rules + dataclass shape
- test_text_utils.py — strip_markdown, tokenise, strip_dialogue,
  make_snippet, ngrams_in_line, read_chapter_drafts
- test_metadata.py — every reader's happy path + missing-file fallback
- test_rules.py — pattern extraction (backtick literal / regex / quoted
  with ban cue), rule_label truncation, end-to-end rule scan
- test_renderer.py — empty-findings message, summary table, per-finding
  recommendation, category-label invariants

Full suite: 1052 tests pass in 2.11s (was 995; +57 from the mirrored
tests, no regressions). ``ruff check .`` is clean.
