+++ tools/analysis/manuscript_checker.py
@@ -1,2201 +1,105 @@
-"""Cross-chapter manuscript checker.
+"""Cross-chapter manuscript checker — backwards-compat shim (Issue #118).
 
-Scans all chapter drafts of a book for prose-quality issues that only surface
-when the whole manuscript is read in one pass. Produces a structured report
-for the `manuscript-checker` skill, which turns it into human-readable
-Markdown with revision recommendations.
+The implementation moved into the ``tools.analysis.manuscript`` package.
+This module stays as a re-export so existing imports keep working:
 
-Detection categories
---------------------
-- **book_rule_violation** — Patterns extracted from the book's CLAUDE.md rules.
-- **simile / character_tell / blocking_tic / sensory / structural /
-  signature_phrase** — Cross-chapter repeated n-grams.
-- **filter_word** — POV-distancing verbs ("felt", "noticed", "saw that") that
-  weaken close-third narration by mediating sensation through the POV head.
-- **adverb_density** — Per-chapter `-ly` adverb ratio. Heavy adverb use is a
-  craft-level red flag independent of repetition.
-- **cliche** — Curated banlist of worn-out fiction phrasings
-  ("blood ran cold", "heart skipped a beat", etc.).
-- **question_as_statement** — Dialogue that starts with an interrogative word
-  but ends with a period instead of a question mark. Usable once as style
-  (McCarthy-style flat delivery), monotonous when systematic.
+- ``from tools.analysis.manuscript_checker import scan_repetitions, render_report``
+- ``from tools.analysis.manuscript_checker import _read_book_rules, ...``
 
-The module is dependency-free (stdlib only) so it can run inside the MCP
-server without extra installs.
+New code should import from ``tools.analysis.manuscript`` directly.
 """
 
 from __future__ import annotations
 
-import re
-from collections import defaultdict
-from dataclasses import dataclass, field, asdict
-from pathlib import Path
-from typing import Any, Iterable
-
-from tools.analysis.callback_validator import verify_callbacks as _verify_callbacks
-
-# Plugin root: two levels up from tools/analysis/
-_PLUGIN_ROOT = Path(__file__).resolve().parents[2]
-
-# ---------------------------------------------------------------------------
-# Book genre detection (stdlib-only frontmatter parse)
-# ---------------------------------------------------------------------------
-
-_FRONTMATTER_BLOCK_RE = re.compile(r"^---\s*\n(.*?)\n---\s*\n", re.DOTALL)
-_GENRES_INLINE_RE = re.compile(r"^\s*genres:\s*\[([^\]]*)\]", re.MULTILINE)
-_GENRES_KEY_RE = re.compile(r"^\s*genres:\s*$", re.MULTILINE)
-_GENRES_ITEM_RE = re.compile(r"^[ \t]+-[ \t]+['\"]?([^'\"#\n]+?)['\"]?\s*$")
-
-
-def _read_book_genres(book_path: Path) -> list[str]:
-    """Parse the genres list from book README.md YAML frontmatter.
-
-    Handles both inline ( genres: ["a", "b"] ) and block ( genres:\\n  - a )
-    formats without requiring a yaml library.
-    """
-    readme = book_path / "README.md"
-    if not readme.is_file():
-        return []
-
-    text = readme.read_text(encoding="utf-8")
-    fm_match = _FRONTMATTER_BLOCK_RE.match(text)
-    if not fm_match:
-        return []
-    frontmatter = fm_match.group(1)
-
-    # Inline list: genres: ["a", "b"] or genres: ['a']
-    inline = _GENRES_INLINE_RE.search(frontmatter)
-    if inline:
-        raw = inline.group(1)
-        return [
-            g.strip().strip("\"'")
-            for g in raw.split(",")
-            if g.strip().strip("\"'")
-        ]
-
-    # Block list: genres:\n  - a\n  - b
-    if _GENRES_KEY_RE.search(frontmatter):
-        genres: list[str] = []
-        in_genres = False
-        for line in frontmatter.splitlines():
-            if re.match(r"^\s*genres:\s*$", line):
-                in_genres = True
-                continue
-            if in_genres:
-                m = _GENRES_ITEM_RE.match(line)
-                if m:
-                    genres.append(m.group(1).strip())
-                elif line.strip() and not line.startswith(" ") and not line.startswith("\t"):
-                    break
-        return genres
-
-    return []
-
-
-_BOOK_CATEGORY_RE = re.compile(
-    r"^\s*book_category:\s*['\"]?(\w+)['\"]?\s*$", re.MULTILINE
+from tools.analysis.manuscript import (  # noqa: F401  (intentional re-exports)
+    BLOCKING_VERBS,
+    BODY_PARTS,
+    DEFAULT_MIN_OCCURRENCES,
+    DEFAULT_NGRAM_SIZES,
+    Finding,
+    Occurrence,
+    SENSORY_TOKENS,
+    STOP_WORDS,
+    STRUCTURAL_HINTS,
+    _classify,
+    _extract_patterns_from_rule,
+    _load_action_verbs,
+    _load_cliche_banlist,
+    _looks_structural,
+    _make_snippet,
+    _ngrams_in_line,
+    _read_allowed_repetitions,
+    _read_book_category,
+    _read_book_genres,
+    _read_book_rules,
+    _read_chapter_drafts,
+    _read_people_profiles,
+    _read_snapshot_threshold,
+    _rule_label,
+    _scan_adverb_density,
+    _scan_anonymization_leak,
+    _scan_book_rules,
+    _scan_callbacks,
+    _scan_cliches,
+    _scan_filter_words,
+    _scan_question_as_statement,
+    _scan_real_people_consistency,
+    _scan_reflective_platitudes,
+    _scan_sentence_repetitions,
+    _scan_snapshots,
+    _scan_tidy_lesson_endings,
+    _scan_timeline_ambiguity,
+    _strip_dialogue,
+    _strip_markdown,
+    _tokenise,
+    render_report,
+    scan_repetitions,
 )
 
-
-def _read_book_category(book_path: Path) -> str:
-    """Parse book_category from book README.md frontmatter. Defaults to 'fiction'."""
-    readme = book_path / "README.md"
-    if not readme.is_file():
-        return "fiction"
-    text = readme.read_text(encoding="utf-8")
-    fm_match = _FRONTMATTER_BLOCK_RE.match(text)
-    if not fm_match:
-        return "fiction"
-    m = _BOOK_CATEGORY_RE.search(fm_match.group(1))
-    return m.group(1).strip() if m else "fiction"
-
-
-# ---------------------------------------------------------------------------
-# People profiles (memoir — Path E #59)
-# ---------------------------------------------------------------------------
-
-_PERSON_NAME_RE = re.compile(
-    r"^\s*name:\s*['\"]?([^'\"\n]+?)['\"]?\s*$", re.MULTILINE
-)
-_PERSON_ANONYMIZATION_RE = re.compile(
-    r"^\s*anonymization:\s*['\"]?([^'\"\n]+?)['\"]?\s*$", re.MULTILINE
-)
-_PERSON_REAL_NAME_RE = re.compile(
-    r"^\s*real_name:\s*['\"]?([^'\"\n]+?)['\"]?\s*$", re.MULTILINE
-)
-
-
-def _read_people_profiles(book_path: Path) -> list[dict[str, str]]:
-    """Read person profiles from book's people/ directory.
-
-    Returns a list of dicts with keys: slug, name, anonymization, real_name.
-    """
-    people_dir = book_path / "people"
-    if not people_dir.is_dir():
-        return []
-
-    people: list[dict[str, str]] = []
-    for person_file in sorted(people_dir.glob("*.md")):
-        if person_file.name == "INDEX.md":
-            continue
-        try:
-            text = person_file.read_text(encoding="utf-8")
-        except OSError:
-            continue
-        fm_match = _FRONTMATTER_BLOCK_RE.match(text)
-        if not fm_match:
-            continue
-        fm = fm_match.group(1)
-        name_m = _PERSON_NAME_RE.search(fm)
-        anon_m = _PERSON_ANONYMIZATION_RE.search(fm)
-        real_m = _PERSON_REAL_NAME_RE.search(fm)
-        people.append({
-            "slug": person_file.stem,
-            "name": name_m.group(1).strip() if name_m else person_file.stem,
-            "anonymization": anon_m.group(1).strip() if anon_m else "none",
-            "real_name": real_m.group(1).strip() if real_m else "",
-        })
-    return people
-
-
-# ---------------------------------------------------------------------------
-# Configuration
-# ---------------------------------------------------------------------------
-
-# Range of n-gram lengths to consider. 4 catches "the ghost of a", 7 catches
-# longer signature constructions like "for a hundred and fifty years" without
-# blowing up the index for very long books.
-DEFAULT_NGRAM_SIZES = (4, 5, 6, 7)
-
-# Minimum occurrences to keep an n-gram. 2 means "appears at least twice".
-DEFAULT_MIN_OCCURRENCES = 2
-
-# Stop-words used for the "skip n-grams that are entirely stop-words" filter.
-# Kept small on purpose — we want to catch "for a hundred and fifty years",
-# which contains stop-words but also distinctive content tokens.
-STOP_WORDS = frozenset(
-    {
-        "a", "an", "the", "and", "or", "but", "if", "then", "of", "to", "in",
-        "on", "at", "by", "for", "with", "from", "into", "onto", "as", "is",
-        "are", "was", "were", "be", "been", "being", "it", "its", "this",
-        "that", "these", "those", "he", "she", "him", "her", "his", "hers",
-        "they", "them", "their", "i", "me", "my", "we", "us", "our", "you",
-        "your", "not", "no", "do", "does", "did", "have", "has", "had", "so",
-        "than", "there", "here", "out", "up", "down", "over", "under", "off",
-        "about", "again", "very", "just",
-    }
-)
-
-# Body-part / face-part vocabulary that, when present in a repeated phrase,
-# strongly suggests a character tell.
-BODY_PARTS = frozenset(
-    {
-        "eye", "eyes", "eyebrow", "eyebrows", "brow", "brows", "lip", "lips",
-        "mouth", "jaw", "jaws", "cheek", "cheeks", "chin", "ear", "ears",
-        "nose", "nostril", "nostrils", "tongue", "throat", "neck", "shoulder",
-        "shoulders", "chest", "back", "spine", "arm", "arms", "elbow",
-        "wrist", "hand", "hands", "finger", "fingers", "thumb", "knuckle",
-        "knuckles", "fist", "fists", "hip", "hips", "leg", "legs", "knee",
-        "knees", "foot", "feet", "toe", "toes", "muscle", "muscles", "vein",
-        "veins", "tendon", "tendons", "head", "skull", "temple", "temples",
-        "scalp", "hair", "face",
-    }
-)
-
-# Verbs that frequently appear in blocking tics (small physical movements
-# repeated as a beat between dialog).
-BLOCKING_VERBS = frozenset(
-    {
-        "opened", "closed", "shut", "blinked", "swallowed", "nodded",
-        "shrugged", "shook", "shifted", "leaned", "looked", "glanced",
-        "stared", "turned", "tilted", "frowned", "smiled", "smirked",
-        "exhaled", "inhaled", "sighed", "shivered", "flinched", "twitched",
-        "jumped", "tightened", "loosened", "pressed", "clenched", "unclenched",
-    }
-)
-
-# Sensory adjectives / nouns. If a repeated phrase contains one of these,
-# bias it toward "sensory repetition".
-SENSORY_TOKENS = frozenset(
-    {
-        "smell", "scent", "stink", "stench", "odour", "odor", "perfume",
-        "taste", "tasted", "tang", "bitter", "sweet", "sour", "salty",
-        "metallic", "coppery", "iron", "rusty",
-        "warm", "cold", "icy", "hot", "burning", "cool",
-        "sound", "noise", "echo", "whisper", "hum", "buzz",
-    }
-)
-
-# Tokens commonly inside structural tics like "the kind of X that Y" or
-# "the first time in X". Used as a hint for the structural category.
-STRUCTURAL_HINTS = frozenset(
-    {"kind", "sort", "type", "first", "last", "only", "way"}
-)
-
-
-# ---------------------------------------------------------------------------
-# Data classes
-# ---------------------------------------------------------------------------
-
-
-@dataclass
-class Occurrence:
-    """A single occurrence of a repeated phrase in the manuscript."""
-
-    chapter: str
-    line: int
-    snippet: str  # ~120 chars of surrounding context, with the phrase intact
-
-
-@dataclass
-class Finding:
-    """A repeated phrase with all its occurrences and a category guess."""
-
-    phrase: str
-    category: str
-    severity: str  # "high" | "medium"
-    count: int
-    occurrences: list[Occurrence] = field(default_factory=list)
-    # Populated only for category == "book_rule_violation": the verbatim rule
-    # from the book's CLAUDE.md that triggered the finding, so the user sees
-    # *why* a phrase was flagged.
-    source_rule: str | None = None
-
-
-# ---------------------------------------------------------------------------
-# Text normalisation
-# ---------------------------------------------------------------------------
-
-_FRONTMATTER_RE = re.compile(r"^---\s*\n.*?\n---\s*\n", re.DOTALL)
-_HEADING_RE = re.compile(r"^\s*#{1,6}\s.*$", re.MULTILINE)
-_HORIZONTAL_RULE_RE = re.compile(r"^\s*[*_-]{3,}\s*$", re.MULTILINE)
-_MD_FORMAT_RE = re.compile(r"[*_`~]")
-# Punctuation to drop for the n-gram identity. We keep apostrophes inside
-# words ("don't") so they aren't split, and we keep "—" as a soft separator.
-_TOKEN_RE = re.compile(r"[A-Za-z][A-Za-z'\u2019]*")
-
-
-def _strip_markdown(text: str) -> str:
-    text = _FRONTMATTER_RE.sub("", text)
-    text = _HEADING_RE.sub("", text)
-    text = _HORIZONTAL_RULE_RE.sub("", text)
-    text = _MD_FORMAT_RE.sub("", text)
-    return text
-
-
-def _tokenise(line: str) -> list[str]:
-    """Lowercase tokens, apostrophes preserved inside words."""
-    return [m.group(0).lower().replace("\u2019", "'") for m in _TOKEN_RE.finditer(line)]
-
-
-# ---------------------------------------------------------------------------
-# Classification
-# ---------------------------------------------------------------------------
-
-# Compiled patterns for category detection on the original (non-normalised)
-# context snippet, so we can spot punctuation cues like "Closed it." or
-# quotation marks.
-_SIMILE_HINT_RE = re.compile(r"\b(like|as if|as though)\b", re.IGNORECASE)
-_AS_X_AS_RE = re.compile(r"\bas\s+\w+\s+as\b", re.IGNORECASE)
-_BLOCKING_PUNCT_RE = re.compile(r"\.\s+\w+(ed|s)?\s+(it|him|her|them)\.")
-
-
-def _classify(phrase: str, occurrences: list[Occurrence]) -> str:
-    """Pick the best category for a repeated phrase."""
-    tokens = phrase.split()
-    token_set = set(tokens)
-    contexts = " ".join(o.snippet for o in occurrences)
-
-    # Similes first — they're the most distinctive.
-    if "like" in token_set or "as" in token_set:
-        if (
-            _SIMILE_HINT_RE.search(contexts)
-            or _AS_X_AS_RE.search(contexts)
-            or "like" in tokens[:2]
-        ):
-            return "simile"
-
-    # Blocking tic: physical micro-action between dialog beats.
-    if token_set & BLOCKING_VERBS:
-        if _BLOCKING_PUNCT_RE.search(contexts) or any(
-            t in token_set for t in ("opened", "closed", "shut")
-        ):
-            return "blocking_tic"
-
-    # Character tell: repeated body-part description.
-    if token_set & BODY_PARTS:
-        return "character_tell"
-
-    # Sensory repetition: same smell/taste/sound description.
-    if token_set & SENSORY_TOKENS:
-        return "sensory"
-
-    # Structural tic: "the kind of X that Y" / "for X years" / "the first time"
-    if token_set & STRUCTURAL_HINTS or _looks_structural(tokens):
-        return "structural"
-
-    return "signature_phrase"
-
-
-def _looks_structural(tokens: list[str]) -> bool:
-    """Heuristic for repeated structural patterns."""
-    if not tokens:
-        return False
-    # "for <number-ish> years/days/months"
-    if tokens[0] == "for" and tokens[-1] in {"years", "year", "days", "months", "weeks"}:
-        return True
-    # "the X of Y" patterns
-    if len(tokens) >= 4 and tokens[0] == "the" and "of" in tokens:
-        return True
-    return False
-
-
-# ---------------------------------------------------------------------------
-# Scanning
-# ---------------------------------------------------------------------------
-
-
-def _make_snippet(line: str, ngram_text: str, max_len: int = 140) -> str:
-    """Return a trimmed snippet around the n-gram match in the original line."""
-    line = line.strip()
-    if len(line) <= max_len:
-        return line
-    idx = line.lower().find(ngram_text)
-    if idx < 0:
-        return line[:max_len].rstrip() + "…"
-    half = max_len // 2
-    start = max(0, idx - half)
-    end = min(len(line), idx + len(ngram_text) + half)
-    snippet = line[start:end].strip()
-    if start > 0:
-        snippet = "…" + snippet
-    if end < len(line):
-        snippet = snippet + "…"
-    return snippet
-
-
-def _ngrams_in_line(
-    tokens: list[str], sizes: Iterable[int]
-) -> list[tuple[int, int, str]]:
-    """Yield (size, start_index, joined_phrase) for every n-gram in the line.
-
-    Skips n-grams that are 100% stop-words — they're noise like "of the and a".
-    """
-    out: list[tuple[int, int, str]] = []
-    for size in sizes:
-        if len(tokens) < size:
-            continue
-        for i in range(len(tokens) - size + 1):
-            window = tokens[i : i + size]
-            if all(t in STOP_WORDS for t in window):
-                continue
-            out.append((size, i, " ".join(window)))
-    return out
-
-
-def _read_chapter_drafts(book_path: Path) -> list[tuple[str, str]]:
-    """Return [(chapter_slug, draft_text), ...] sorted by chapter folder name."""
-    chapters_dir = book_path / "chapters"
-    if not chapters_dir.is_dir():
-        return []
-    drafts: list[tuple[str, str]] = []
-    for chapter_dir in sorted(chapters_dir.iterdir()):
-        if not chapter_dir.is_dir():
-            continue
-        draft_path = chapter_dir / "draft.md"
-        if not draft_path.exists():
-            continue
-        try:
-            text = draft_path.read_text(encoding="utf-8")
-        except OSError:
-            continue
-        drafts.append((chapter_dir.name, text))
-    return drafts
-
-
-# ---------------------------------------------------------------------------
-# Per-book CLAUDE.md rule scan
-# ---------------------------------------------------------------------------
-
-# Match "## Rules" heading through to the next "## " heading or EOF.
-_RULES_SECTION_RE = re.compile(
-    r"^##\s+Rules\s*$(.*?)(?=^##\s+\S|\Z)",
-    re.MULTILINE | re.DOTALL,
-)
-
-# Match a markdown list item that spans until the next blank line, next list
-# item, or section end. This preserves multi-sentence rules written on one
-# logical bullet (continued across wrapped lines).
-_RULE_BULLET_RE = re.compile(
-    r"^-\s+(?P<body>.+?)(?=^-\s+|^\s*$|^<!--|\Z)",
-    re.MULTILINE | re.DOTALL,
-)
-
-# Comment markers inside the Rules section — stripped before bullet parsing.
-_COMMENT_RE = re.compile(r"<!--.*?-->", re.DOTALL)
-
-# Backtick-wrapped content. We split on these to distinguish regex hints from
-# plain-literal tokens.
-_BACKTICK_CONTENT_RE = re.compile(r"`([^`\n]+)`")
-
-# Double-quoted phrases ≥3 chars of content. Deliberately excludes short words
-# like "a" or "ok" which produce noisy false positives.
-_QUOTED_CONTENT_RE = re.compile(r'"([^"\n]{3,})"')
-
-# Characters that strongly suggest a backtick-wrapped string is intended as a
-# regex rather than a literal substring.
-_REGEX_HINT_CHARS = set("|()[]\\^$?+*{}")
-
-# Cue keywords that mark a rule as containing bannable quoted phrases. Without
-# a cue, quoted strings are treated as examples, not patterns.
-_BAN_CUE_RE = re.compile(
-    r"\b(banned|ban|avoid|never|don[\u2019']?t\s+use|do\s+not\s+use|limit|no\s+\w+)\b",
-    re.IGNORECASE,
-)
-
-
-def _read_book_rules(book_path: Path) -> list[str]:
-    """Extract rule text entries from a book's CLAUDE.md.
-
-    Returns one string per bullet item found under the ``## Rules`` section,
-    including entries inside the ``<!-- RULES:START --> ... <!-- RULES:END -->``
-    block and any static entries listed above it. Comment markers are stripped
-    before bullet parsing so they don't break list items.
-
-    Returns an empty list when CLAUDE.md is missing or has no Rules section.
-    """
-    claudemd = book_path / "CLAUDE.md"
-    if not claudemd.is_file():
-        return []
-    try:
-        text = claudemd.read_text(encoding="utf-8")
-    except OSError:
-        return []
-
-    match = _RULES_SECTION_RE.search(text)
-    if not match:
-        return []
-    section = _COMMENT_RE.sub("", match.group(1))
-
-    rules: list[str] = []
-    for m in _RULE_BULLET_RE.finditer(section):
-        body = m.group("body").strip()
-        # Collapse internal whitespace (from line wrapping) to single spaces.
-        body = re.sub(r"\s+", " ", body)
-        if body:
-            rules.append(body)
-    return rules
-
-
-def _extract_patterns_from_rule(rule: str) -> list[tuple[str, re.Pattern[str]]]:
-    """Extract scannable patterns from a single rule text.
-
-    Returns a list of ``(display_label, compiled_regex)`` tuples. Heuristic
-    extraction — stdlib only:
-
-    1. Backtick-wrapped strings are always extracted. If the content contains
-       regex metacharacters it's compiled as a regex, otherwise as a literal
-       substring. Whitespace inside the backticks is preserved so the user
-       can encode word-boundary intent (e.g. `` ` thing ` ``).
-    2. Double-quoted phrases are extracted *only* when the rule contains a
-       ban cue (``banned``, ``avoid``, ``never``, ``don't use``, ``do not
-       use``, ``ban``, ``limit``, ``no X``). This prevents positive rewrite
-       examples from being interpreted as patterns to ban.
-    3. Italics (``*foo*``) are intentionally ignored — they're used for
-       narrative examples, not scannable bans.
-    4. Malformed regex strings are skipped rather than raising.
-    """
-    patterns: list[tuple[str, re.Pattern[str]]] = []
-    seen: set[str] = set()
-
-    def _add(label: str, compiled: re.Pattern[str]) -> None:
-        key = compiled.pattern.lower()
-        if key in seen:
-            return
-        seen.add(key)
-        patterns.append((label, compiled))
-
-    # Backticks: always extracted.
-    for m in _BACKTICK_CONTENT_RE.finditer(rule):
-        raw = m.group(1)
-        inner = raw.strip()
-        if len(inner) < 2:
-            continue
-        if any(c in _REGEX_HINT_CHARS for c in inner):
-            try:
-                _add(inner, re.compile(raw, re.IGNORECASE))
-            except re.error:
-                continue
-        else:
-            _add(inner, re.compile(re.escape(raw), re.IGNORECASE))
-
-    # Quoted phrases: only if the rule contains a ban cue. Require at least
-    # 6 chars after strip — shorter quoted strings are usually concept
-    # references (e.g. the title `"thing"`) rather than bannable phrases.
-    if _BAN_CUE_RE.search(rule):
-        for m in _QUOTED_CONTENT_RE.finditer(rule):
-            raw = m.group(1).strip()
-            if len(raw) < 6 or raw.lower() in STOP_WORDS:
-                continue
-            _add(raw, re.compile(re.escape(raw), re.IGNORECASE))
-
-    return patterns
-
-
-def _rule_label(rule: str, max_len: int = 80) -> str:
-    """Short display label for a rule — typically the bold'd title prefix."""
-    # Prefer text inside the first **bold** span if present.
-    bold = re.match(r"\*\*(?P<title>[^*]+)\*\*", rule)
-    if bold:
-        title = bold.group("title").strip()
-    else:
-        title = rule
-    title = re.sub(r"\s+", " ", title)
-    if len(title) > max_len:
-        title = title[: max_len - 1].rstrip() + "…"
-    return title
-
-
-def _scan_book_rules(book_path: Path) -> list[Finding]:
-    """Scan chapter drafts for violations of rules in the book's CLAUDE.md.
-
-    Produces one ``Finding`` per (rule × pattern) that has at least one hit,
-    with severity always ``"high"`` because user-authored rules override
-    frequency thresholds.
-    """
-    rules = _read_book_rules(book_path)
-    if not rules:
-        return []
-    drafts = _read_chapter_drafts(book_path)
-    if not drafts:
-        return []
-
-    findings: list[Finding] = []
-    for rule in rules:
-        patterns = _extract_patterns_from_rule(rule)
-        if not patterns:
-            continue
-        rule_label = _rule_label(rule)
-        # Aggregate all patterns of a rule into a single finding. Dedupe
-        # per (chapter, line) so overlapping patterns within the same rule
-        # don't produce double-counted occurrences.
-        seen_positions: set[tuple[str, int]] = set()
-        occurrences: list[Occurrence] = []
-        matched_labels: dict[str, None] = {}  # insertion-ordered unique set
-        for display, pattern in patterns:
-            pattern_hit = False
-            for chapter_slug, raw_text in drafts:
-                cleaned = _strip_markdown(raw_text)
-                for line_no, line in enumerate(cleaned.splitlines(), start=1):
-                    stripped = line.strip()
-                    if not stripped:
-                        continue
-                    for m in pattern.finditer(stripped):
-                        key = (chapter_slug, line_no)
-                        if key in seen_positions:
-                            continue
-                        seen_positions.add(key)
-                        snippet = _make_snippet(stripped, m.group(0).lower())
-                        occurrences.append(
-                            Occurrence(chapter=chapter_slug, line=line_no, snippet=snippet)
-                        )
-                        pattern_hit = True
-            if pattern_hit:
-                matched_labels[display] = None
-        if not occurrences:
-            continue
-        phrase = " / ".join(matched_labels) if matched_labels else rule_label
-        findings.append(
-            Finding(
-                phrase=phrase,
-                category="book_rule_violation",
-                severity="high",
-                count=len(occurrences),
-                occurrences=sorted(occurrences, key=lambda o: (o.chapter, o.line)),
-                source_rule=rule_label,
-            )
-        )
-    return findings
-
-
-# ---------------------------------------------------------------------------
-# Filter words — POV distancing verbs
-# ---------------------------------------------------------------------------
-
-# Verbs that mediate sensation through the POV character's head rather than
-# letting the reader experience it directly. Overuse is a close-third
-# weakness. Patterns use word boundaries and, where necessary, require a
-# disambiguating follower (e.g. "saw that" vs. just "saw") to avoid flagging
-# legitimate uses.
-FILTER_WORD_PATTERNS: tuple[tuple[str, re.Pattern[str]], ...] = (
-    ("felt",        re.compile(r"\b(?:felt)\b", re.IGNORECASE)),
-    ("noticed",     re.compile(r"\bnoticed\b", re.IGNORECASE)),
-    ("saw that",    re.compile(r"\bsaw\s+(?:that|how|the\s+way)\b", re.IGNORECASE)),
-    ("heard that",  re.compile(r"\bheard\s+(?:that|how|the)\b", re.IGNORECASE)),
-    ("seemed",      re.compile(r"\bseemed\b", re.IGNORECASE)),
-    ("appeared",    re.compile(r"\bappeared\s+to\b", re.IGNORECASE)),
-    ("realized",    re.compile(r"\brealized\b", re.IGNORECASE)),
-    ("wondered",    re.compile(r"\bwondered\b", re.IGNORECASE)),
-    ("watched",     re.compile(r"\bwatched\b", re.IGNORECASE)),
-    ("observed",    re.compile(r"\bobserved\b", re.IGNORECASE)),
-    ("thought that", re.compile(r"\bthought\s+(?:that|of)\b", re.IGNORECASE)),
-    ("decided",     re.compile(r"\bdecided\b", re.IGNORECASE)),
-    ("knew that",   re.compile(r"\bknew\s+(?:that|how)\b", re.IGNORECASE)),
-    ("remembered",  re.compile(r"\bremembered\b", re.IGNORECASE)),
-    ("sensed",      re.compile(r"\bsensed\b", re.IGNORECASE)),
-)
-
-# Per-chapter thresholds. Below medium_threshold is acceptable. Between
-# medium_threshold and high_threshold is medium severity; at or above
-# high_threshold is high severity. Expressed as counts-per-1000-words so
-# short chapters aren't over-penalised.
-FILTER_WORD_MEDIUM_PER_1K = 3.0
-FILTER_WORD_HIGH_PER_1K = 6.0
-
-
-def _scan_filter_words(book_path: Path) -> list[Finding]:
-    """Flag POV filter-word overuse per chapter.
-
-    Produces one Finding per chapter that exceeds the medium threshold,
-    listing the top filter-word hits plus total density.
-    """
-    drafts = _read_chapter_drafts(book_path)
-    if not drafts:
-        return []
-
-    findings: list[Finding] = []
-    for chapter_slug, raw_text in drafts:
-        cleaned = _strip_markdown(raw_text)
-        word_count = len(_tokenise(cleaned))
-        if word_count < 200:
-            # Too short to draw a meaningful density conclusion.
-            continue
-
-        occurrences: list[Occurrence] = []
-        per_word_counts: dict[str, int] = defaultdict(int)
-        for line_no, line in enumerate(cleaned.splitlines(), start=1):
-            stripped = line.strip()
-            if not stripped:
-                continue
-            # Skip dialogue-only lines: filter words inside quotes are
-            # character speech, not narration.
-            narration = _strip_dialogue(stripped)
-            if not narration.strip():
-                continue
-            for label, pattern in FILTER_WORD_PATTERNS:
-                for m in pattern.finditer(narration):
-                    per_word_counts[label] += 1
-                    # Snippet from the original line so the reader sees the
-                    # full context (including any surrounding dialogue).
-                    occurrences.append(
-                        Occurrence(
-                            chapter=chapter_slug,
-                            line=line_no,
-                            snippet=_make_snippet(stripped, m.group(0).lower()),
-                        )
-                    )
-
-        if not occurrences:
-            continue
-
-        density = (len(occurrences) / word_count) * 1000.0
-        if density < FILTER_WORD_MEDIUM_PER_1K:
-            continue
-        severity = "high" if density >= FILTER_WORD_HIGH_PER_1K else "medium"
-
-        # Build phrase label showing top offenders.
-        top = sorted(per_word_counts.items(), key=lambda kv: -kv[1])[:3]
-        top_str = ", ".join(f"{word}×{n}" for word, n in top)
-        phrase = f"{chapter_slug}: {top_str} ({density:.1f}/1k words)"
-
-        findings.append(
-            Finding(
-                phrase=phrase,
-                category="filter_word",
-                severity=severity,
-                count=len(occurrences),
-                occurrences=occurrences[:20],  # cap per chapter
-            )
-        )
-    return findings
-
-
-# ---------------------------------------------------------------------------
-# Adverb density — `-ly` adverbs per 1000 words
-# ---------------------------------------------------------------------------
-
-# `-ly` words that are NOT adverbs or are unavoidable (pronouns, nouns,
-# common function words). Excluded from the density count to keep the signal
-# meaningful.
-_LY_EXCLUSIONS = frozenset(
-    {
-        # Not adverbs
-        "only", "family", "belly", "jelly", "rally", "folly", "holly", "silly",
-        "bully", "lily", "ally", "really",  # "really" is an adverb but so
-        # common (often in dialogue) that flagging it adds noise
-        "early", "lovely", "lonely", "lively", "friendly", "deadly",
-        "ugly", "holy", "homely", "ghastly", "ghostly", "gnarly", "scholarly",
-        "timely", "costly", "oily", "hilly", "jolly", "chilly", "wooly",
-        "woolly", "manly", "knightly", "kingly", "queenly", "princely",
-        # Proper/place names that often end in -ly (kept small)
-        "italy",
-    }
-)
-
-# Thresholds per 1000 words.
-ADVERB_MEDIUM_PER_1K = 8.0
-ADVERB_HIGH_PER_1K = 14.0
-
-_LY_WORD_RE = re.compile(r"\b([a-z]+ly)\b", re.IGNORECASE)
-
-
-def _scan_adverb_density(book_path: Path) -> list[Finding]:
-    """Flag chapters with heavy `-ly` adverb density.
-
-    One Finding per flagged chapter with the top adverbs and density metric.
-    """
-    drafts = _read_chapter_drafts(book_path)
-    if not drafts:
-        return []
-
-    findings: list[Finding] = []
-    for chapter_slug, raw_text in drafts:
-        cleaned = _strip_markdown(raw_text)
-        tokens = _tokenise(cleaned)
-        word_count = len(tokens)
-        if word_count < 200:
-            continue
-
-        adverb_counts: dict[str, int] = defaultdict(int)
-        occurrences: list[Occurrence] = []
-        for line_no, line in enumerate(cleaned.splitlines(), start=1):
-            stripped = line.strip()
-            if not stripped:
-                continue
-            # Count adverbs in narration only — dialogue adverbs are the
-            # character's speech, not the narrator's voice.
-            narration = _strip_dialogue(stripped)
-            for m in _LY_WORD_RE.finditer(narration):
-                word = m.group(1).lower()
-                if word in _LY_EXCLUSIONS:
-                    continue
-                # Skip 3-letter "-ly" like "sly" / "fly" — they aren't adverbs.
-                if len(word) < 5:
-                    continue
-                adverb_counts[word] += 1
-                if len(occurrences) < 20:
-                    occurrences.append(
-                        Occurrence(
-                            chapter=chapter_slug,
-                            line=line_no,
-                            snippet=_make_snippet(stripped, word),
-                        )
-                    )
-
-        total = sum(adverb_counts.values())
-        if total == 0:
-            continue
-        density = (total / word_count) * 1000.0
-        if density < ADVERB_MEDIUM_PER_1K:
-            continue
-        severity = "high" if density >= ADVERB_HIGH_PER_1K else "medium"
-
-        top = sorted(adverb_counts.items(), key=lambda kv: -kv[1])[:5]
-        top_str = ", ".join(f"{w}×{n}" for w, n in top)
-        phrase = f"{chapter_slug}: {top_str} ({density:.1f}/1k words, {total} total)"
-
-        findings.append(
-            Finding(
-                phrase=phrase,
-                category="adverb_density",
-                severity=severity,
-                count=total,
-                occurrences=occurrences,
-            )
-        )
-    return findings
-
-
-# ---------------------------------------------------------------------------
-# Cliché detection — curated banlist
-# ---------------------------------------------------------------------------
-
-# Curated list of the worst-offender fiction clichés. Kept deliberately short
-# — better to catch the 30 phrases that are unambiguously stale than to
-# false-positive on borderline imagery. Each entry compiled case-insensitive,
-# word-bounded where needed.
-CLICHE_PHRASES: tuple[str, ...] = (
-    # Cardiovascular clichés
-    "blood ran cold",
-    "heart skipped a beat",
-    "heart sank",
-    "heart pounded in his chest",
-    "heart pounded in her chest",
-    "blood boiled",
-    "pulse quickened",
-    # Ocular / facial clichés
-    "eyes widened in horror",
-    "eyes narrowed",
-    "rolled her eyes",
-    "rolled his eyes",
-    "locked eyes",
-    "eyes met across the room",
-    # Time / cosmic clichés
-    "time stood still",
-    "time seemed to slow",
-    "the world fell away",
-    "everything went black",
-    "an eternity passed",
-    # Breath / voice clichés
-    "breath caught in her throat",
-    "breath caught in his throat",
-    "lump in his throat",
-    "lump in her throat",
-    "barely above a whisper",
-    # Weather / atmosphere
-    "it was a dark and stormy night",
-    "a chill ran down his spine",
-    "a chill ran down her spine",
-    "hair stood on end",
-    "hair on the back of his neck stood",
-    "hair on the back of her neck stood",
-    # Misc narrative
-    "little did he know",
-    "little did she know",
-    "only time would tell",
-    "sight for sore eyes",
-    "needle in a haystack",
-    "calm before the storm",
-)
-
-
-_ENTRY_RE = re.compile(
-    r"^\s*-\s+(.+?)(?:\s+\(severity:\s+(high|medium)\))?(?:\s+—.*)?$",
-    re.IGNORECASE,
-)
-
-
-def _load_cliche_banlist(
-    plugin_root: Path,
-    genres: list[str] | None = None,
-) -> list[tuple[str, str]]:
-    """Load the cliché banlist from reference/craft/cliche-banlist.md.
-
-    Returns a list of (phrase, severity) tuples. Falls back to the hardcoded
-    CLICHE_PHRASES constant (all severity 'high') when the file is missing.
-    Each genre in `genres` merges its cliche-banlist-{genre}.md on top;
-    files that don't exist are silently skipped.
-    """
-    craft_dir = plugin_root / "reference" / "craft"
-
-    def _parse_file(path: Path) -> list[tuple[str, str]]:
-        if not path.is_file():
-            return []
-        entries: list[tuple[str, str]] = []
-        for line in path.read_text(encoding="utf-8").splitlines():
-            m = _ENTRY_RE.match(line)
-            if not m:
-                continue
-            phrase = m.group(1).strip()
-            severity = (m.group(2) or "high").lower()
-            if phrase and len(phrase) > 2:
-                entries.append((phrase, severity))
-        return entries
-
-    base = _parse_file(craft_dir / "cliche-banlist.md")
-    if not base:
-        base = [(p, "high") for p in CLICHE_PHRASES]
-
-    if genres:
-        seen = {p for p, _ in base}
-        for genre_slug in genres:
-            for phrase, severity in _parse_file(
-                craft_dir / f"cliche-banlist-{genre_slug}.md"
-            ):
-                if phrase not in seen:
-                    base.append((phrase, severity))
-                    seen.add(phrase)
-
-    return base
-
-
-def _compile_cliche_patterns() -> tuple[tuple[str, re.Pattern[str]], ...]:
-    return tuple(
-        (phrase, re.compile(r"\b" + re.escape(phrase) + r"\b", re.IGNORECASE))
-        for phrase in CLICHE_PHRASES
-    )
-
-
-_CLICHE_PATTERNS = _compile_cliche_patterns()
-
-
-def _scan_cliches(
-    book_path: Path,
-    plugin_root: Path | None = None,
-    genres: list[str] | None = None,
-) -> list[Finding]:
-    """Flag hits from the curated cliché banlist.
-
-    Loads patterns from the versioned reference file when available; falls back
-    to the hardcoded CLICHE_PHRASES constant. One Finding per unique phrase.
-    """
-    drafts = _read_chapter_drafts(book_path)
-    if not drafts:
-        return []
-
-    root = plugin_root if plugin_root is not None else _PLUGIN_ROOT
-    banlist = _load_cliche_banlist(root, genres=genres)
-    patterns = [
-        (phrase, severity, re.compile(r"\b" + re.escape(phrase) + r"\b", re.IGNORECASE))
-        for phrase, severity in banlist
-    ]
-
-    findings: list[Finding] = []
-    for phrase, severity, pattern in patterns:
-        occurrences: list[Occurrence] = []
-        for chapter_slug, raw_text in drafts:
-            cleaned = _strip_markdown(raw_text)
-            for line_no, line in enumerate(cleaned.splitlines(), start=1):
-                stripped = line.strip()
-                if not stripped:
-                    continue
-                for m in pattern.finditer(stripped):
-                    occurrences.append(
-                        Occurrence(
-                            chapter=chapter_slug,
-                            line=line_no,
-                            snippet=_make_snippet(stripped, m.group(0).lower()),
-                        )
-                    )
-        if occurrences:
-            findings.append(
-                Finding(
-                    phrase=phrase,
-                    category="cliche",
-                    severity=severity,
-                    count=len(occurrences),
-                    occurrences=sorted(occurrences, key=lambda o: (o.chapter, o.line)),
-                )
-            )
-    return findings
-
-
-# ---------------------------------------------------------------------------
-# Question-as-statement — dialogue punctuation anomaly
-# ---------------------------------------------------------------------------
-
-# Interrogative openers. First token of the dialogue. Covers wh-questions and
-# yes/no aux-verb questions.
-_QUESTION_OPENERS = frozenset(
-    {
-        # wh-questions
-        "who", "what", "where", "when", "why", "how", "which", "whose",
-        # aux-verb yes/no questions
-        "do", "does", "did", "is", "are", "was", "were", "am",
-        "can", "could", "will", "would", "shall", "should", "may", "might", "must",
-        "have", "has", "had",
-        # Shortened forms occasionally used ("n't" prefixed onto aux is
-        # tough to catch without a tokenizer; skip for now)
-    }
-)
-
-# Matches text inside straight or curly double-quoted spans. Non-greedy;
-# must have at least 2 chars of content.
-_DIALOGUE_RE = re.compile(
-    r'(?:"([^"\n]{2,}?)"|\u201C([^\u201C\u201D\n]{2,}?)\u201D)'
-)
-
-
-def _strip_dialogue(line: str) -> str:
-    """Remove dialogue (quoted text) from a line, leaving narration only.
-
-    Used by filter-word and adverb-density scans so we only count narrator
-    tics, not character speech.
-    """
-    # Replace each quoted span with a space so surrounding tokens don't merge.
-    return _DIALOGUE_RE.sub(" ", line)
-
-
-def _scan_question_as_statement(book_path: Path) -> list[Finding]:
-    """Flag dialogue that starts with a question word but ends with a period.
-
-    McCarthy-style flat delivery is legitimate once in a while, but becomes
-    monotonous when it's the default. The detector reports every hit so the
-    author can judge which ones to convert to question marks and which ones
-    to keep as flat demands (recommendation: pair a kept one with a
-    narrative beat like "It was a demand, not a question.").
-    """
-    drafts = _read_chapter_drafts(book_path)
-    if not drafts:
-        return []
-
-    occurrences: list[Occurrence] = []
-    for chapter_slug, raw_text in drafts:
-        cleaned = _strip_markdown(raw_text)
-        for line_no, line in enumerate(cleaned.splitlines(), start=1):
-            stripped = line.strip()
-            if not stripped:
-                continue
-            for m in _DIALOGUE_RE.finditer(stripped):
-                dialogue = (m.group(1) or m.group(2) or "").strip()
-                if len(dialogue) < 2:
-                    continue
-                # Must end with a period. Exclude `?`, `!`, `…`, trailing
-                # em-dash (interrupted) and `--` (ASCII interrupted).
-                last = dialogue[-1]
-                if last != ".":
-                    continue
-                # Ellipsis like "What..." — trailing dots are not a single
-                # period, skip.
-                if dialogue.endswith("..") or dialogue.endswith("\u2026"):
-                    continue
-                # Must START with an interrogative token.
-                first_tokens = _tokenise(dialogue)
-                if not first_tokens:
-                    continue
-                first = first_tokens[0]
-                # Contractions ("don't", "can't", "isn't") collapse to
-                # "don", "can", "isn" after tokenisation — we check the
-                # leading alpha run and treat those as their auxiliary.
-                if first in _QUESTION_OPENERS or first in {
-                    "don", "doesn", "didn", "isn", "aren", "wasn",
-                    "weren", "can", "couldn", "won", "wouldn", "shouldn",
-                    "hasn", "haven", "hadn",
-                }:
-                    snippet = _make_snippet(stripped, m.group(0).lower())
-                    occurrences.append(
-                        Occurrence(
-                            chapter=chapter_slug,
-                            line=line_no,
-                            snippet=snippet,
-                        )
-                    )
-
-    if not occurrences:
-        return []
-
-    # Single aggregated finding: high severity if 5+ hits, medium otherwise.
-    severity = "high" if len(occurrences) >= 5 else "medium"
-    return [
-        Finding(
-            phrase='Dialogue Q-word ending with "." instead of "?"',
-            category="question_as_statement",
-            severity=severity,
-            count=len(occurrences),
-            occurrences=sorted(occurrences, key=lambda o: (o.chapter, o.line))[:40],
-        )
-    ]
-
-
-# ---------------------------------------------------------------------------
-# Allowed repetitions — book-level exclusion list
-# ---------------------------------------------------------------------------
-
-_ALLOWED_REPS_RE = re.compile(r"^##\s+Allowed Repetitions\s*$", re.IGNORECASE)
-
-
-def _read_allowed_repetitions(book_path: Path) -> frozenset[str]:
-    """Parse ## Allowed Repetitions from book CLAUDE.md.
-
-    Returns a frozenset of lowercased, tokenised phrases that should be
-    excluded from the sentence-level repetition scan.
-    """
-    claudemd = book_path / "CLAUDE.md"
-    if not claudemd.is_file():
-        return frozenset()
-
-    allowed: set[str] = set()
-    in_section = False
-    for line in claudemd.read_text(encoding="utf-8").splitlines():
-        if _ALLOWED_REPS_RE.match(line):
-            in_section = True
-            continue
-        if in_section:
-            if line.startswith("#"):
-                break
-            stripped = line.strip().lstrip("- ").strip()
-            if stripped:
-                normalized = " ".join(_tokenise(stripped))
-                if normalized:
-                    allowed.add(normalized)
-    return frozenset(allowed)
-
-
-# ---------------------------------------------------------------------------
-# Sentence-level repetition detector (8-15 word n-grams)
-# ---------------------------------------------------------------------------
-
-DEFAULT_SENTENCE_NGRAM_SIZES = tuple(range(8, 16))
-
-
-def _scan_sentence_repetitions(
-    book_path: Path,
-    min_length: int = 8,
-    max_length: int = 15,
-    min_occurrences: int = 2,
-) -> list[Finding]:
-    """Detect repeated whole-sentence-level phrases (8-15 word n-grams).
-
-    The existing 4-7 word detector catches short constructions; this pass
-    targets longer units — the most egregious AI tell when reused verbatim.
-    Phrases listed under ## Allowed Repetitions in the book CLAUDE.md are
-    excluded.
-    """
-    drafts = _read_chapter_drafts(book_path)
-    if not drafts:
-        return []
-
-    allowed = _read_allowed_repetitions(book_path)
-    sizes = tuple(range(min_length, max_length + 1))
-    index: dict[str, list[Occurrence]] = defaultdict(list)
-
-    for chapter_slug, raw_text in drafts:
-        cleaned = _strip_markdown(raw_text)
-        for line_no, original_line in enumerate(cleaned.splitlines(), start=1):
-            stripped = original_line.strip()
-            if not stripped:
-                continue
-            tokens = _tokenise(stripped)
-            if len(tokens) < min_length:
-                continue
-            for _size, _i, phrase in _ngrams_in_line(tokens, sizes):
-                index[phrase].append(
-                    Occurrence(
-                        chapter=chapter_slug,
-                        line=line_no,
-                        snippet=_make_snippet(stripped, phrase),
-                    )
-                )
-
-    findings: list[Finding] = []
-    for phrase, occs in index.items():
-        if len(occs) < min_occurrences:
-            continue
-        # Skip if any allowed repetition contains this phrase as a sub-phrase
-        if any(phrase in allowed_phrase for allowed_phrase in allowed):
-            continue
-        findings.append(
-            Finding(
-                phrase=phrase,
-                category="sentence_repetition",
-                severity="high",
-                count=len(occs),
-                occurrences=sorted(occs, key=lambda o: (o.chapter, o.line)),
-            )
-        )
-
-    return findings
-
-
-# ---------------------------------------------------------------------------
-# Snapshot detector — static description blocks without movement or dialog
-# ---------------------------------------------------------------------------
-
-_SNAPSHOT_THRESHOLD_DEFAULT = 5
-_SNAPSHOT_THRESHOLD_RE = re.compile(
-    r"^\s*-\s+snapshot_threshold:\s*(\d+)", re.MULTILINE
-)
-_LINTER_CONFIG_RE = re.compile(r"^##\s+Linter Config\s*$", re.MULTILINE)
-
-# Fallback action verbs when the reference file is missing.
-_ACTION_VERBS_FALLBACK = frozenset(
-    {
-        "walk", "run", "move", "step", "reach", "grab", "turn", "look", "sit",
-        "stand", "open", "close", "push", "pull", "say", "ask", "tell", "call",
-        "enter", "leave", "fall", "rise", "jump", "catch", "throw", "pull",
-        "lift", "carry", "take", "give", "find", "decide", "realize", "notice",
-    }
-)
-
-# Split on sentence-ending punctuation followed by whitespace.
-# Deliberately lenient: ellipsis (...) and single-char abbreviations are
-# handled by requiring at least one letter before the end marker.
-_SENTENCE_SPLIT_RE = re.compile(r"(?<=[a-zA-Z\d][.!?])\s+")
-_DIALOG_CHAR_RE = re.compile(r'["“”]')
-
-
-def _read_snapshot_threshold(book_path: Path) -> int:
-    """Read per-book snapshot threshold from ## Linter Config in CLAUDE.md."""
-    claudemd = book_path / "CLAUDE.md"
-    if not claudemd.is_file():
-        return _SNAPSHOT_THRESHOLD_DEFAULT
-    text = claudemd.read_text(encoding="utf-8")
-    # Only look inside the ## Linter Config section
-    lc = _LINTER_CONFIG_RE.search(text)
-    if not lc:
-        return _SNAPSHOT_THRESHOLD_DEFAULT
-    section = text[lc.start():]
-    # Stop at next heading
-    next_heading = re.search(r"^##\s", section[3:], re.MULTILINE)
-    if next_heading:
-        section = section[: next_heading.start() + 3]
-    m = _SNAPSHOT_THRESHOLD_RE.search(section)
-    if m:
-        return max(2, int(m.group(1)))
-    return _SNAPSHOT_THRESHOLD_DEFAULT
-
-
-def _load_action_verbs(plugin_root: Path) -> frozenset[str]:
-    """Load the action verb list from reference/craft/action-verbs.md.
-
-    Returns a frozenset of lowercased base verb forms. Falls back to a minimal
-    hardcoded set if the file is missing.
-    """
-    path = plugin_root / "reference" / "craft" / "action-verbs.md"
-    if not path.is_file():
-        return _ACTION_VERBS_FALLBACK
-    verbs: set[str] = set()
-    for line in path.read_text(encoding="utf-8").splitlines():
-        stripped = line.strip().lstrip("- ").strip()
-        if stripped and not stripped.startswith("#") and len(stripped) >= 2:
-            verbs.add(stripped.lower())
-    return frozenset(verbs) if verbs else _ACTION_VERBS_FALLBACK
-
-
-def _sentence_has_action(sentence: str, action_verbs: frozenset[str]) -> bool:
-    """Return True if the sentence contains at least one action verb."""
-    tokens = set(_tokenise(sentence))
-    for verb in action_verbs:
-        if verb in tokens:
-            return True
-        # Common inflections
-        if f"{verb}s" in tokens or f"{verb}ed" in tokens or f"{verb}ing" in tokens:
-            return True
-        # Drop trailing -e before -ing: take → taking, move → moving
-        if verb.endswith("e") and f"{verb[:-1]}ing" in tokens:
-            return True
-        # Doubling final consonant: run → running, sit → sitting
-        if len(verb) >= 3 and verb[-1] == verb[-2] and f"{verb}ing" in tokens:
-            return True
-    return False
-
-
-def _scan_snapshots(
-    book_path: Path,
-    plugin_root: Path | None = None,
-) -> list[Finding]:
-    """Flag static description blocks with no action verbs and no dialog.
-
-    Reads the threshold from ## Linter Config → snapshot_threshold (default 5).
-    One Finding per flagged block; severity is always medium.
-    """
-    drafts = _read_chapter_drafts(book_path)
-    if not drafts:
-        return []
-
-    threshold = _read_snapshot_threshold(book_path)
-    root = plugin_root if plugin_root is not None else _PLUGIN_ROOT
-    action_verbs = _load_action_verbs(root)
-
-    findings: list[Finding] = []
-
-    for chapter_slug, raw_text in drafts:
-        cleaned = _strip_markdown(raw_text)
-        lines = cleaned.splitlines()
-
-        # Collect paragraphs (blank-line-separated blocks of text).
-        paragraphs: list[tuple[int, str]] = []  # (start_line, text)
-        para_start = 0
-        para_lines: list[str] = []
-        for line_no, line in enumerate(lines, start=1):
-            if line.strip():
-                if not para_lines:
-                    para_start = line_no
-                para_lines.append(line)
-            else:
-                if para_lines:
-                    paragraphs.append((para_start, " ".join(para_lines)))
-                    para_lines = []
-        if para_lines:
-            paragraphs.append((para_start, " ".join(para_lines)))
-
-        for para_start_line, para_text in paragraphs:
-            sentences = [
-                s.strip()
-                for s in _SENTENCE_SPLIT_RE.split(para_text)
-                if s.strip()
-            ]
-            if len(sentences) < threshold:
-                continue
-
-            # Walk through sentences counting consecutive descriptive ones.
-            streak: list[str] = []
-            streak_start = para_start_line
-
-            def _flush_streak(streak: list[str], start: int) -> None:
-                if len(streak) >= threshold:
-                    snippet = " ".join(streak)[:200].rstrip()
-                    if len(" ".join(streak)) > 200:
-                        snippet += "…"
-                    findings.append(
-                        Finding(
-                            phrase=f"snapshot:{chapter_slug}:{start}",
-                            category="snapshot",
-                            severity="medium",
-                            count=len(streak),
-                            occurrences=[
-                                Occurrence(
-                                    chapter=chapter_slug,
-                                    line=start,
-                                    snippet=snippet,
-                                )
-                            ],
-                        )
-                    )
-
-            for sent in sentences:
-                if not sent:
-                    continue
-                if _DIALOG_CHAR_RE.search(sent) or _sentence_has_action(sent, action_verbs):
-                    _flush_streak(streak, streak_start)
-                    streak = []
-                else:
-                    if not streak:
-                        streak_start = para_start_line
-                    streak.append(sent)
-
-            _flush_streak(streak, streak_start)
-
-    return findings
-
-
-# ---------------------------------------------------------------------------
-# Callback register scanner
-# ---------------------------------------------------------------------------
-
-_CALLBACK_DEFERRED_SILENCE = 10  # chapters of silence before deferred becomes a finding
-
-
-def _scan_callbacks(book_path: Path) -> list[Finding]:
-    """Surface broken callback promises as manuscript findings.
-
-    Reads the book's CLAUDE.md, runs verify_callbacks, and converts:
-    - ``potentially_dropped`` entries → high-severity ``callback_dropped`` findings
-    - ``deferred`` entries with chapters_since > threshold → medium-severity
-      ``callback_deferred`` findings
-
-    Satisfied callbacks and recent-deferred callbacks produce no findings.
-    """
-    claudemd_path = book_path / "CLAUDE.md"
-    if not claudemd_path.exists():
-        return []
-
-    claudemd_text = claudemd_path.read_text(encoding="utf-8")
-    result = _verify_callbacks(book_path, claudemd_text)
-
-    findings: list[Finding] = []
-
-    for entry in result.get("potentially_dropped", []):
-        warning = entry.get("warning", "no appearance found")
-        occ = Occurrence(chapter="CLAUDE.md", line=0, snippet=warning)
-        findings.append(Finding(
-            phrase=entry["name"],
-            category="callback_dropped",
-            severity="high",
-            count=entry.get("chapters_since", 0),
-            occurrences=[occ],
-        ))
-
-    for entry in result.get("deferred", []):
-        chapters_since = entry.get("chapters_since", 0)
-        if chapters_since <= _CALLBACK_DEFERRED_SILENCE:
-            continue
-        snippet = f"not appeared in {chapters_since} drafted chapters"
-        occ = Occurrence(chapter="CLAUDE.md", line=0, snippet=snippet)
-        findings.append(Finding(
-            phrase=entry["name"],
-            category="callback_deferred",
-            severity="medium",
-            count=chapters_since,
-            occurrences=[occ],
-        ))
-
-    return findings
-
-
-# ---------------------------------------------------------------------------
-# Memoir-specific checks (Phase 3, Issue #61)
-# ---------------------------------------------------------------------------
-
-# Temporal hand-waving phrases that leave the reader unanchored in time.
-_TEMPORAL_VAGUE_PATTERNS: tuple[tuple[str, re.Pattern[str]], ...] = (
-    ("some time later",  re.compile(r"\bsome\s+time\s+later\b", re.IGNORECASE)),
-    ("at some point",    re.compile(r"\bat\s+some\s+point\b", re.IGNORECASE)),
-    ("one day",          re.compile(r"\bone\s+day\b", re.IGNORECASE)),
-    ("one night",        re.compile(r"\bone\s+night\b", re.IGNORECASE)),
-    ("eventually",       re.compile(r"\beventually\b", re.IGNORECASE)),
-    ("after a while",    re.compile(r"\bafter\s+a\s+while\b", re.IGNORECASE)),
-    ("a while later",    re.compile(r"\ba\s+while\s+later\b", re.IGNORECASE)),
-    ("years later",      re.compile(r"\byears\s+later\b", re.IGNORECASE)),
-    ("years earlier",    re.compile(r"\byears\s+earlier\b", re.IGNORECASE)),
-    ("years before",     re.compile(r"\byears\s+before\b", re.IGNORECASE)),
-    ("years ago",        re.compile(r"\byears\s+ago\b", re.IGNORECASE)),
-    ("a long time ago",  re.compile(r"\ba\s+long\s+time\s+ago\b", re.IGNORECASE)),
-    ("back then",        re.compile(r"\bback\s+then\b", re.IGNORECASE)),
-    ("in those days",    re.compile(r"\bin\s+those\s+days\b", re.IGNORECASE)),
-    ("at that time",     re.compile(r"\bat\s+that\s+time\b", re.IGNORECASE)),
-    ("around that time", re.compile(r"\baround\s+that\s+time\b", re.IGNORECASE)),
-    ("sometime later",   re.compile(r"\bsometime\s+later\b", re.IGNORECASE)),
-    ("not long after",   re.compile(r"\bnot\s+long\s+after\b", re.IGNORECASE)),
-    ("before long",      re.compile(r"\bbefore\s+long\b", re.IGNORECASE)),
-)
-
-TIMELINE_AMBIGUITY_MEDIUM_PER_1K = 3.0
-TIMELINE_AMBIGUITY_HIGH_PER_1K = 6.0
-
-
-def _scan_timeline_ambiguity(book_path: Path) -> list[Finding]:
-    """Memoir: flag chapters with excessive temporal hand-waving.
-
-    A memoir scene that never anchors when it happened loses credibility. This
-    check flags chapters where temporal vagueness density is high — not single
-    uses (which are legitimate shorthand), but systemic avoidance of specific
-    time anchoring.
-    """
-    drafts = _read_chapter_drafts(book_path)
-    if not drafts:
-        return []
-
-    findings: list[Finding] = []
-    for chapter_slug, raw_text in drafts:
-        cleaned = _strip_markdown(raw_text)
-        word_count = len(_tokenise(cleaned))
-        if word_count < 200:
-            continue
-
-        occurrences: list[Occurrence] = []
-        per_pattern_counts: dict[str, int] = defaultdict(int)
-
-        for line_no, line in enumerate(cleaned.splitlines(), start=1):
-            stripped = line.strip()
-            if not stripped:
-                continue
-            for label, pattern in _TEMPORAL_VAGUE_PATTERNS:
-                for m in pattern.finditer(stripped):
-                    per_pattern_counts[label] += 1
-                    occurrences.append(
-                        Occurrence(
-                            chapter=chapter_slug,
-                            line=line_no,
-                            snippet=_make_snippet(stripped, m.group(0).lower()),
-                        )
-                    )
-
-        if not occurrences:
-            continue
-
-        density = (len(occurrences) / word_count) * 1000.0
-        if density < TIMELINE_AMBIGUITY_MEDIUM_PER_1K:
-            continue
-
-        severity = "high" if density >= TIMELINE_AMBIGUITY_HIGH_PER_1K else "medium"
-        top = sorted(per_pattern_counts.items(), key=lambda kv: -kv[1])[:3]
-        top_str = ", ".join(f"{label}×{n}" for label, n in top)
-        phrase = f"{chapter_slug}: {top_str} ({density:.1f}/1k words)"
-
-        findings.append(
-            Finding(
-                phrase=phrase,
-                category="timeline_ambiguity",
-                severity=severity,
-                count=len(occurrences),
-                occurrences=occurrences[:20],
-            )
-        )
-    return findings
-
-
-# Retrospective commentary patterns that collapse scene into TED talk.
-_REFLECTIVE_PLATITUDE_PATTERNS: tuple[tuple[str, re.Pattern[str]], ...] = (
-    ("looking back",         re.compile(r"\blooking\s+back\b", re.IGNORECASE)),
-    ("in retrospect",        re.compile(r"\bin\s+retrospect\b", re.IGNORECASE)),
-    ("in hindsight",         re.compile(r"\bin\s+hindsight\b", re.IGNORECASE)),
-    ("with hindsight",       re.compile(r"\bwith\s+hindsight\b", re.IGNORECASE)),
-    ("I now realize",        re.compile(r"\bI\s+now\s+realiz", re.IGNORECASE)),
-    ("I now understand",     re.compile(r"\bI\s+now\s+understand\b", re.IGNORECASE)),
-    ("I now know",           re.compile(r"\bI\s+now\s+know\b", re.IGNORECASE)),
-    ("I realize now",        re.compile(r"\bI\s+realiz\w*\s+now\b", re.IGNORECASE)),
-    ("I understand now",     re.compile(r"\bI\s+understand\s+now\b", re.IGNORECASE)),
-    ("what I learned",       re.compile(r"\bwhat\s+I\s+learned\b", re.IGNORECASE)),
-    ("I came to realize",    re.compile(r"\bI\s+came\s+to\s+realiz", re.IGNORECASE)),
-    ("I came to understand", re.compile(r"\bI\s+came\s+to\s+understand\b", re.IGNORECASE)),
-    ("I would later",        re.compile(r"\bI\s+would\s+later\b", re.IGNORECASE)),
-    ("it taught me",         re.compile(r"\bit\s+taught\s+me\b", re.IGNORECASE)),
-    ("taught me that",       re.compile(r"\btaught\s+me\s+that\b", re.IGNORECASE)),
-    ("I had come to",        re.compile(r"\bI\s+had\s+come\s+to\b", re.IGNORECASE)),
-    ("the lesson was",       re.compile(r"\bthe\s+lesson\s+was\b", re.IGNORECASE)),
-)
-
-PLATITUDE_MEDIUM_THRESHOLD = 2  # per chapter (absolute count)
-PLATITUDE_HIGH_THRESHOLD = 3
-
-
-def _scan_reflective_platitudes(book_path: Path) -> list[Finding]:
-    """Memoir: flag chapters heavy with retrospective lesson-summary language.
-
-    Some reflective narration is memoir's lifeblood. But dense retrospective
-    commentary ("looking back", "in hindsight", "what I learned") turns a lived
-    scene into a TED talk. Flag chapters where the count crosses the threshold.
-    """
-    drafts = _read_chapter_drafts(book_path)
-    if not drafts:
-        return []
-
-    findings: list[Finding] = []
-    for chapter_slug, raw_text in drafts:
-        cleaned = _strip_markdown(raw_text)
-
-        occurrences: list[Occurrence] = []
-        per_pattern_counts: dict[str, int] = defaultdict(int)
-
-        for line_no, line in enumerate(cleaned.splitlines(), start=1):
-            stripped = line.strip()
-            if not stripped:
-                continue
-            for label, pattern in _REFLECTIVE_PLATITUDE_PATTERNS:
-                for m in pattern.finditer(stripped):
-                    per_pattern_counts[label] += 1
-                    occurrences.append(
-                        Occurrence(
-                            chapter=chapter_slug,
-                            line=line_no,
-                            snippet=_make_snippet(stripped, m.group(0).lower()),
-                        )
-                    )
-
-        if len(occurrences) < PLATITUDE_MEDIUM_THRESHOLD:
-            continue
-
-        severity = "high" if len(occurrences) >= PLATITUDE_HIGH_THRESHOLD else "medium"
-        top = sorted(per_pattern_counts.items(), key=lambda kv: -kv[1])[:3]
-        top_str = ", ".join(f"{label}×{n}" for label, n in top)
-        phrase = f"{chapter_slug}: {top_str} ({len(occurrences)} hits)"
-
-        findings.append(
-            Finding(
-                phrase=phrase,
-                category="reflective_platitude",
-                severity=severity,
-                count=len(occurrences),
-                occurrences=occurrences[:20],
-            )
-        )
-    return findings
-
-
-# Lesson-summary patterns in chapter endings.
-_LESSON_ENDING_PATTERNS: tuple[re.Pattern[str], ...] = (
-    re.compile(r"\bI\s+(?:had\s+)?learned\b", re.IGNORECASE),
-    re.compile(r"\btaught\s+me\b", re.IGNORECASE),
-    re.compile(r"\bmade\s+me\s+realiz", re.IGNORECASE),
-    re.compile(r"\bhelped\s+me\s+(?:realiz|understand|see)\b", re.IGNORECASE),
-    re.compile(r"\bshowed\s+me\s+that\b", re.IGNORECASE),
-    re.compile(r"\bmade\s+me\s+understand\b", re.IGNORECASE),
-    re.compile(r"\bI\s+(?:had\s+)?come\s+to\s+understand\b", re.IGNORECASE),
-    re.compile(r"\bI\s+now\s+(?:understood|knew)\b", re.IGNORECASE),
-    re.compile(r"\bwhat\s+I\s+came\s+away\s+with\b", re.IGNORECASE),
-    re.compile(r"\bchanged\s+(?:the\s+way\s+I|me\s+forever|everything)\b", re.IGNORECASE),
-    re.compile(r"\bI\s+would\s+never\s+(?:forget|be\s+the\s+same)\b", re.IGNORECASE),
-    re.compile(r"\bthe\s+lesson\s+(?:I|was|here)\b", re.IGNORECASE),
-)
-
-
-def _last_paragraph(text: str) -> str:
-    """Return the last non-empty paragraph of a text block."""
-    paragraphs = [p.strip() for p in text.split("\n\n") if p.strip()]
-    return paragraphs[-1] if paragraphs else ""
-
-
-def _scan_tidy_lesson_endings(book_path: Path) -> list[Finding]:
-    """Memoir: flag chapters whose final paragraph closes on a moral or lesson summary.
-
-    Great memoir endings leave the reader in the moment — not in a retrospective
-    summary. A chapter ending with "and that is what taught me X" explains rather
-    than renders. Flag chapters where the last paragraph has multiple lesson-cues.
-    """
-    drafts = _read_chapter_drafts(book_path)
-    if not drafts:
-        return []
-
-    findings: list[Finding] = []
-    for chapter_slug, raw_text in drafts:
-        cleaned = _strip_markdown(raw_text)
-        last_para = _last_paragraph(cleaned)
-        if not last_para or len(last_para.split()) < 10:
-            continue
-
-        hits: list[str] = []
-        for pattern in _LESSON_ENDING_PATTERNS:
-            for m in pattern.finditer(last_para):
-                hits.append(m.group(0))
-
-        if len(hits) < 2:
-            continue
-
-        severity = "high" if len(hits) >= 3 else "medium"
-        snippet = last_para[:200].rstrip()
-        if len(last_para) > 200:
-            snippet += "…"
-
-        findings.append(
-            Finding(
-                phrase=f"{chapter_slug}: chapter ends on lesson-summary language",
-                category="tidy_lesson_ending",
-                severity=severity,
-                count=len(hits),
-                occurrences=[
-                    Occurrence(chapter=chapter_slug, line=0, snippet=snippet)
-                ],
-            )
-        )
-    return findings
-
-
-def _scan_anonymization_leak(book_path: Path) -> list[Finding]:
-    """Memoir: flag draft chapters that contain a person's real name despite anonymization.
-
-    A single occurrence is high severity — this is a potential privacy or
-    defamation leak that must be caught before publication.
-    """
-    people = _read_people_profiles(book_path)
-    anon_people = [
-        p for p in people
-        if p["anonymization"] != "none" and p["real_name"]
-    ]
-    if not anon_people:
-        return []
-
-    drafts = _read_chapter_drafts(book_path)
-    if not drafts:
-        return []
-
-    findings: list[Finding] = []
-    for person in anon_people:
-        real_name = person["real_name"]
-        display_name = person["name"]
-        pattern = re.compile(r"\b" + re.escape(real_name) + r"\b", re.IGNORECASE)
-
-        occurrences: list[Occurrence] = []
-        for chapter_slug, raw_text in drafts:
-            cleaned = _strip_markdown(raw_text)
-            for line_no, line in enumerate(cleaned.splitlines(), start=1):
-                stripped = line.strip()
-                if not stripped:
-                    continue
-                for m in pattern.finditer(stripped):
-                    occurrences.append(
-                        Occurrence(
-                            chapter=chapter_slug,
-                            line=line_no,
-                            snippet=_make_snippet(stripped, m.group(0).lower()),
-                        )
-                    )
-
-        if occurrences:
-            findings.append(
-                Finding(
-                    phrase=f'real name "{real_name}" (pseudonym: "{display_name}")',
-                    category="anonymization_leak",
-                    severity="high",
-                    count=len(occurrences),
-                    occurrences=sorted(occurrences, key=lambda o: (o.chapter, o.line)),
-                )
-            )
-    return findings
-
-
-def _scan_real_people_consistency(book_path: Path) -> list[Finding]:
-    """Memoir: flag inconsistent capitalisation/form of a person's display name across chapters.
-
-    If "Anna" appears as "anna" in one chapter and "Anna" in another, that is
-    an inconsistency the author should resolve — especially for pseudonymous
-    characters where case variation can break anonymization conventions.
-    """
-    people = _read_people_profiles(book_path)
-    if not people:
-        return []
-
-    drafts = _read_chapter_drafts(book_path)
-    if not drafts:
-        return []
-
-    findings: list[Finding] = []
-
-    for person in people:
-        display_name = person["name"]
-        if not display_name or len(display_name) < 2:
-            continue
-
-        # Case-sensitive search: collect distinct forms found across chapters.
-        forms_found: dict[str, list[Occurrence]] = defaultdict(list)
-        name_pattern = re.compile(r"\b" + re.escape(display_name) + r"\b", re.IGNORECASE)
-
-        for chapter_slug, raw_text in drafts:
-            cleaned = _strip_markdown(raw_text)
-            for line_no, line in enumerate(cleaned.splitlines(), start=1):
-                stripped = line.strip()
-                if not stripped:
-                    continue
-                for m in name_pattern.finditer(stripped):
-                    matched_form = m.group(0)  # preserve original case
-                    forms_found[matched_form].append(
-                        Occurrence(
-                            chapter=chapter_slug,
-                            line=line_no,
-                            snippet=_make_snippet(stripped, matched_form.lower()),
-                        )
-                    )
-
-        if len(forms_found) <= 1:
-            continue
-
-        forms_str = ", ".join(f'"{f}"' for f in sorted(forms_found.keys()))
-        all_occurrences = [
-            occ for occs in forms_found.values() for occ in occs
-        ]
-        findings.append(
-            Finding(
-                phrase=f'"{display_name}": multiple forms found — {forms_str}',
-                category="real_people_consistency",
-                severity="medium",
-                count=len(forms_found),
-                occurrences=sorted(
-                    all_occurrences, key=lambda o: (o.chapter, o.line)
-                )[:10],
-            )
-        )
-    return findings
-
-
-# ---------------------------------------------------------------------------
-# Scanning
-# ---------------------------------------------------------------------------
-
-
-def scan_repetitions(
-    book_path: Path,
-    ngram_sizes: Iterable[int] = DEFAULT_NGRAM_SIZES,
-    min_occurrences: int = DEFAULT_MIN_OCCURRENCES,
-    max_findings_per_category: int | None = None,
-    plugin_root: Path | None = None,
-    book_category: str | None = None,
-) -> dict[str, Any]:
-    """Scan all chapter drafts of a book for repeated phrases.
-
-    Returns a dict with `findings` (list[Finding-as-dict]), grouped counts,
-    and basic metadata. The caller — typically the MCP tool — turns this into
-    a Markdown report.
-    """
-    drafts = _read_chapter_drafts(book_path)
-    if not drafts:
-        return {
-            "book_path": str(book_path),
-            "chapters_scanned": 0,
-            "findings": [],
-            "summary": {},
-        }
-
-    # Map: phrase -> list[Occurrence]
-    index: dict[str, list[Occurrence]] = defaultdict(list)
-    sizes = tuple(ngram_sizes)
-
-    for chapter_slug, raw_text in drafts:
-        cleaned = _strip_markdown(raw_text)
-        for line_no, original_line in enumerate(cleaned.splitlines(), start=1):
-            stripped = original_line.strip()
-            if not stripped:
-                continue
-            tokens = _tokenise(stripped)
-            if len(tokens) < min(sizes):
-                continue
-            for _size, _i, phrase in _ngrams_in_line(tokens, sizes):
-                snippet = _make_snippet(stripped, phrase)
-                index[phrase].append(
-                    Occurrence(chapter=chapter_slug, line=line_no, snippet=snippet)
-                )
-
-    # Per-length thresholds. Shorter n-grams need MORE occurrences to be
-    # considered noteworthy, otherwise the report drowns in 4-grams that are
-    # just common English connective tissue.
-    length_thresholds = {
-        4: max(min_occurrences, 5),
-        5: max(min_occurrences, 3),
-        6: max(min_occurrences, 2),
-        7: max(min_occurrences, 2),
-    }
-
-    # Build findings, dropping sub-phrases that a longer accepted phrase
-    # already explains. Tolerance of 1 catches the case where the shorter
-    # phrase appears once outside the longer match.
-    findings: list[Finding] = []
-    sorted_phrases = sorted(index.keys(), key=lambda p: (-len(p.split()), p))
-    seen_long: list[tuple[str, int]] = []  # (phrase, count) of accepted longer phrases
-
-    for phrase in sorted_phrases:
-        occs = index[phrase]
-        size = len(phrase.split())
-        threshold = length_thresholds.get(size, min_occurrences)
-        if len(occs) < threshold:
-            continue
-        # Skip if a longer accepted phrase fully contains this one with a
-        # comparable occurrence count (within +/- 1).
-        if any(
-            phrase in longer and abs(len(occs) - longer_count) <= 1
-            for longer, longer_count in seen_long
-        ):
-            continue
-        category = _classify(phrase, occs)
-        severity = "high" if len(occs) >= 4 else "medium"
-        findings.append(
-            Finding(
-                phrase=phrase,
-                category=category,
-                severity=severity,
-                count=len(occs),
-                occurrences=sorted(occs, key=lambda o: (o.chapter, o.line)),
-            )
-        )
-        seen_long.append((phrase, len(occs)))
-
-    # Merge in per-book CLAUDE.md rule violations. These are high-severity
-    # by definition and ignore the n-gram frequency thresholds above.
-    findings.extend(_scan_book_rules(book_path))
-    # Merge the craft-level checks (filter words, adverb density, clichés,
-    # question-as-statement punctuation, sentence-level repetitions).
-    findings.extend(_scan_filter_words(book_path))
-    findings.extend(_scan_adverb_density(book_path))
-    book_genres = _read_book_genres(book_path)
-    findings.extend(
-        _scan_cliches(
-            book_path,
-            plugin_root=plugin_root,
-            genres=book_genres or None,
-        )
-    )
-    findings.extend(_scan_question_as_statement(book_path))
-    findings.extend(_scan_sentence_repetitions(book_path))
-    findings.extend(_scan_snapshots(book_path, plugin_root=plugin_root))
-    findings.extend(_scan_callbacks(book_path))
-
-    # Memoir-specific checks — run only when book_category is memoir.
-    if book_category is None:
-        book_category = _read_book_category(book_path)
-    if book_category == "memoir":
-        findings.extend(_scan_anonymization_leak(book_path))
-        findings.extend(_scan_tidy_lesson_endings(book_path))
-        findings.extend(_scan_reflective_platitudes(book_path))
-        findings.extend(_scan_timeline_ambiguity(book_path))
-        findings.extend(_scan_real_people_consistency(book_path))
-
-    # Sort order priority: book_rule_violation first (user-authored rules
-    # override everything), then anonymization_leak (privacy-critical),
-    # then clichés (always bad), then the rest by severity.
-    # Within same bucket: severity desc, count desc, phrase asc.
-    category_rank = {
-        "book_rule_violation": 0,
-        "anonymization_leak": 1,
-        "cliche": 2,
-    }
-    severity_rank = {"high": 0, "medium": 1}
-    findings.sort(
-        key=lambda f: (
-            category_rank.get(f.category, 2),
-            severity_rank[f.severity],
-            -f.count,
-            f.phrase,
-        )
-    )
-
-    if max_findings_per_category is not None:
-        per_cat: dict[str, int] = defaultdict(int)
-        capped: list[Finding] = []
-        for f in findings:
-            if per_cat[f.category] >= max_findings_per_category:
-                continue
-            per_cat[f.category] += 1
-            capped.append(f)
-        findings = capped
-
-    # Summary
-    summary: dict[str, dict[str, int]] = defaultdict(lambda: {"high": 0, "medium": 0})
-    for f in findings:
-        summary[f.category][f.severity] += 1
-
-    return {
-        "book_path": str(book_path),
-        "chapters_scanned": len(drafts),
-        "findings": [_finding_to_dict(f) for f in findings],
-        "summary": {k: dict(v) for k, v in summary.items()},
-    }
-
-
-def _finding_to_dict(f: Finding) -> dict[str, Any]:
-    d = asdict(f)
-    return d
-
-
-# ---------------------------------------------------------------------------
-# Markdown report
-# ---------------------------------------------------------------------------
-
-CATEGORY_LABELS = {
-    "book_rule_violation": "Book Rule Violations",
-    "cliche": "Clichés",
-    "question_as_statement": "Dialogue Punctuation (Q-word + period)",
-    "filter_word": "POV Filter Words",
-    "adverb_density": "Adverb Density (per Chapter)",
-    "sentence_repetition": "Sentence-Level Repetitions (8-15 words)",
-    "snapshot": "Snapshot Blocks (static description, no movement)",
-    "callback_dropped": "Dropped Callbacks (promised threads, no follow-through)",
-    "callback_deferred": "Deferred Callbacks (long-silent registered threads)",
-    # Memoir-specific (Phase 3, #61)
-    "anonymization_leak": "Anonymization Leaks (real name in manuscript)",
-    "tidy_lesson_ending": "Tidy-Lesson Endings (chapter closes on a moral)",
-    "reflective_platitude": "Reflective Platitude Density (retrospective commentary)",
-    "timeline_ambiguity": "Timeline Ambiguity (temporal hand-waving)",
-    "real_people_consistency": "Real-People Name Inconsistency",
-    # N-gram repetition categories
-    "simile": "Similes & Metaphors",
-    "blocking_tic": "Blocking Tics",
-    "character_tell": "Character Tells",
-    "sensory": "Sensory Repetitions",
-    "structural": "Structural Tics",
-    "signature_phrase": "Signature Phrases",
-}
-
-CATEGORY_ORDER = [
-    "book_rule_violation",
-    "anonymization_leak",
-    "cliche",
-    "question_as_statement",
-    "filter_word",
-    "adverb_density",
-    "sentence_repetition",
-    "snapshot",
-    "callback_dropped",
-    "callback_deferred",
-    # Memoir-specific
-    "tidy_lesson_ending",
-    "reflective_platitude",
-    "timeline_ambiguity",
-    "real_people_consistency",
-    # N-gram repetition categories
-    "simile",
-    "character_tell",
-    "blocking_tic",
-    "structural",
-    "sensory",
-    "signature_phrase",
+__all__ = [
+    # Public API
+    "Finding",
+    "Occurrence",
+    "render_report",
+    "scan_repetitions",
+    # Constants
+    "BLOCKING_VERBS",
+    "BODY_PARTS",
+    "DEFAULT_MIN_OCCURRENCES",
+    "DEFAULT_NGRAM_SIZES",
+    "SENSORY_TOKENS",
+    "STOP_WORDS",
+    "STRUCTURAL_HINTS",
+    # Private helpers (re-exported for tests + skill-side imports)
+    "_classify",
+    "_extract_patterns_from_rule",
+    "_load_action_verbs",
+    "_load_cliche_banlist",
+    "_looks_structural",
+    "_make_snippet",
+    "_ngrams_in_line",
+    "_read_allowed_repetitions",
+    "_read_book_category",
+    "_read_book_genres",
+    "_read_book_rules",
+    "_read_chapter_drafts",
+    "_read_people_profiles",
+    "_read_snapshot_threshold",
+    "_rule_label",
+    "_scan_adverb_density",
+    "_scan_anonymization_leak",
+    "_scan_book_rules",
+    "_scan_callbacks",
+    "_scan_cliches",
+    "_scan_filter_words",
+    "_scan_question_as_statement",
+    "_scan_real_people_consistency",
+    "_scan_reflective_platitudes",
+    "_scan_sentence_repetitions",
+    "_scan_snapshots",
+    "_scan_tidy_lesson_endings",
+    "_scan_timeline_ambiguity",
+    "_strip_dialogue",
+    "_strip_markdown",
+    "_tokenise",
 ]
-
-
-def render_report(scan_result: dict[str, Any]) -> str:
-    """Turn a scan result into a human-readable Markdown report."""
-    findings = scan_result.get("findings", [])
-    chapters_scanned = scan_result.get("chapters_scanned", 0)
-    summary = scan_result.get("summary", {})
-
-    lines: list[str] = []
-    lines.append("# Repetition Report")
-    lines.append("")
-    lines.append(f"**Chapters scanned:** {chapters_scanned}")
-    lines.append(f"**Findings:** {len(findings)}")
-    lines.append("")
-
-    if not findings:
-        lines.append("No cross-chapter repetitions detected with the current thresholds.")
-        lines.append("")
-        return "\n".join(lines)
-
-    # Summary table
-    lines.append("## Summary")
-    lines.append("")
-    lines.append("| Category | High (4+) | Medium (2-3) |")
-    lines.append("|---|---:|---:|")
-    for cat in CATEGORY_ORDER:
-        if cat not in summary:
-            continue
-        s = summary[cat]
-        lines.append(
-            f"| {CATEGORY_LABELS[cat]} | {s.get('high', 0)} | {s.get('medium', 0)} |"
-        )
-    lines.append("")
-
-    # Group findings by category
-    by_category: dict[str, list[dict[str, Any]]] = defaultdict(list)
-    for f in findings:
-        by_category[f["category"]].append(f)
-
-    for cat in CATEGORY_ORDER:
-        items = by_category.get(cat, [])
-        if not items:
-            continue
-        lines.append(f"## {CATEGORY_LABELS[cat]}")
-        lines.append("")
-        for f in items:
-            severity_marker = "**HIGH**" if f["severity"] == "high" else "MEDIUM"
-            lines.append(f"### `{f['phrase']}` — {f['count']}× ({severity_marker})")
-            lines.append("")
-            if f.get("source_rule"):
-                lines.append(f"> **Rule:** {f['source_rule']}")
-                lines.append("")
-            for occ in f["occurrences"]:
-                lines.append(f"- **{occ['chapter']}** (line {occ['line']}): {occ['snippet']}")
-            lines.append("")
-            lines.append(_recommendation_for(f))
-            lines.append("")
-    return "\n".join(lines)
-
-
-def _recommendation_for(finding: dict[str, Any]) -> str:
-    """Short, category-aware revision recommendation."""
-    cat = finding["category"]
-    count = finding["count"]
-    if cat == "book_rule_violation":
-        return (
-            f"_Recommendation:_ This violates a rule from the book's CLAUDE.md. "
-            f"Rewrite per the user-authored guidance above — all {count} "
-            f"occurrence{'s' if count != 1 else ''} should be revised unless "
-            f"the rule explicitly allows an exception."
-        )
-    if cat == "cliche":
-        return (
-            f"_Recommendation:_ \"{finding['phrase']}\" is a worn-out fiction "
-            f"cliché. Replace every occurrence with imagery specific to this "
-            f"scene's POV, stakes, and sensory palette. If you must keep one, "
-            f"make it ironic or subvert it."
-        )
-    if cat == "question_as_statement":
-        return (
-            "_Recommendation:_ A single flat-delivery question reads as a "
-            "stylistic choice (think McCarthy). At this density it reads as "
-            "a missing keystroke. Two fixes: **(A)** convert to a real "
-            "question mark — most dialogue wants this. **(B)** keep the "
-            "period and pair it with a narrative beat that tells the reader "
-            "the delivery is deliberate, e.g.:\n\n"
-            "> \"Who?\"\n"
-            "> It was a demand, not a question.\n\n"
-            "Pick (A) as the default. Reserve (B) for moments where the "
-            "flatness is load-bearing."
-        )
-    if cat == "filter_word":
-        return (
-            "_Recommendation:_ Filter words mediate sensation through the POV "
-            "character's head — \"she felt the cold\" instead of \"the cold "
-            "bit through her coat\". In close-third, they weaken immersion. "
-            "Rewrite most hits by dropping the filter verb and letting the "
-            "sensation act directly on the scene. Some are load-bearing "
-            "(internal realisation, dream logic); keep those."
-        )
-    if cat == "adverb_density":
-        return (
-            "_Recommendation:_ Heavy `-ly` adverb use usually signals weak "
-            "verb choice: \"walked slowly\" → \"trudged\", \"said quietly\" → "
-            "\"murmured\". Not every adverb is wrong — but when density is "
-            "this high, at least half are propping up verbs that could stand "
-            "on their own. Strip them and see what survives."
-        )
-    if cat == "simile":
-        return (
-            f"_Recommendation:_ Keep the strongest occurrence (usually the first) "
-            f"and rewrite the other {count - 1} with fresh imagery rooted in the "
-            f"current scene's senses."
-        )
-    if cat == "character_tell":
-        return (
-            "_Recommendation:_ A repeated body-part tell becomes invisible after "
-            "the second use. Keep one or two, then vary the physical signal — "
-            "a different body part, an action, or a verbal beat."
-        )
-    if cat == "blocking_tic":
-        return (
-            f"_Recommendation:_ Blocking beats lose impact when reused. Replace "
-            f"{count - 1} occurrences with action that advances the scene or "
-            f"reveals subtext."
-        )
-    if cat == "structural":
-        return (
-            f"_Recommendation:_ A structural tic ({count}×) trains the reader to "
-            f"see the seams. Recast the weaker instances with different syntax."
-        )
-    if cat == "sensory":
-        return (
-            f"_Recommendation:_ Same sensory description in {count} places — "
-            f"vary at least {count - 1} of them so each scene has its own texture."
-        )
-    if cat == "snapshot":
-        count = finding["count"]
-        return (
-            f"_Recommendation:_ {count} consecutive descriptive sentences with no "
-            f"action and no dialog — the scene becomes a photograph. Fix options: "
-            f"**(A)** insert one beat of character action (reaching for something, "
-            f"shifting weight, a micro-gesture that shows state of mind); "
-            f"**(B)** cut the block to 2-3 sentences of pure setting and trust the "
-            f"reader to fill in the rest; "
-            f"**(C)** if description is intentional (lyrical pause, aftermath beat), "
-            f"raise the per-book threshold in `## Linter Config → snapshot_threshold`."
-        )
-    if cat == "sentence_repetition":
-        return (
-            f"_Recommendation:_ This {len(finding['phrase'].split())}-word sentence "
-            f"recurs {count}× — the loudest AI tell in the book. Do not fix by swapping "
-            f"one word; the rhythm and structure are what the reader recognises. Rewrite "
-            f"the emotional beat entirely: different body signal, different duration, "
-            f"different syntax. If it is a deliberate motif, add it to the book's "
-            f"## Allowed Repetitions section in CLAUDE.md."
-        )
-    if cat == "callback_dropped":
-        return (
-            f"_Recommendation:_ '{finding['phrase']}' is registered in the Callback "
-            f"Register with a hard deadline or must-not-forget flag that has been "
-            f"breached. Either (A) plant the callback in the appropriate chapter now, "
-            f"(B) update the register entry to reflect the new plan, or "
-            f"(C) remove the callback if the thread was intentionally dropped."
-        )
-    if cat == "callback_deferred":
-        return (
-            f"_Recommendation:_ '{finding['phrase']}' has been silent for {count} "
-            f"chapters. Decide whether to plant it soon or remove it from the register."
-        )
-    if cat == "anonymization_leak":
-        return (
-            "_Recommendation:_ A person's real name appears in the manuscript despite "
-            "their profile being marked as anonymized. Replace every occurrence with "
-            "the pseudonym (or a relationship-term like 'my colleague') before the "
-            "manuscript leaves the author's desk. This is a pre-publication blocker."
-        )
-    if cat == "tidy_lesson_ending":
-        return (
-            "_Recommendation:_ The chapter ends by explaining what the experience "
-            "meant rather than letting the moment speak. Cut the lesson language and "
-            "close on a concrete detail — an image, a gesture, a line of dialogue. "
-            "The reader draws the meaning; the author renders the scene."
-        )
-    if cat == "reflective_platitude":
-        return (
-            "_Recommendation:_ Dense retrospective commentary ('looking back', "
-            "'in hindsight', 'what I learned') collapses scene into TED talk. "
-            "Memoir earns its reflection by first rendering the experience fully. "
-            "Cut or push the commentary phrases to a single moment of narrating-self "
-            "intrusion; let the rest of the chapter stay in the experiencing self."
-        )
-    if cat == "timeline_ambiguity":
-        return (
-            "_Recommendation:_ Temporal hand-waving ('at some point', 'eventually', "
-            "'one day') leaves the reader unanchored. Memoir credibility depends on "
-            "specificity: a season, a year, a day-of-week, a life stage. Replace at "
-            "least the chapter-opening time anchor with something concrete. A few "
-            "vague transitions inside a chapter are fine; flagging density means too "
-            "many in a row with no specific date anywhere in the chapter."
-        )
-    if cat == "real_people_consistency":
-        return (
-            "_Recommendation:_ The same person is referred to by different name forms "
-            "across chapters. Pick the canonical form (the pseudonym, or the first-name "
-            "only, or the full pseudonym) and apply it consistently throughout. "
-            "Inconsistency confuses readers and can partially undermine anonymization."
-        )
-    return (
-        f"_Recommendation:_ Decide which occurrence is most necessary; cut or "
-        f"rewrite the other {count - 1}."
-    )
